# swagger.json
{"openapi":"3.0.0","paths":{"/api/version":{"get":{"operationId":"AppController_getVersion","summary":"getVersion","description":"Get the app version","parameters":[],"responses":{"200":{"description":"Return current version","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VersionRes"}}}}},"tags":["System"]}},"/api/healthz":{"get":{"operationId":"AppController_healthz","summary":"healthz","description":"For metrics server health checking","parameters":[],"responses":{"200":{"description":"Return OK","content":{"application/json":{"example":{"data":"OK"}}}}},"tags":["System"]}},"/api/stay":{"post":{"operationId":"StayController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStayDto"}}}},"responses":{"201":{"description":""}}}},"/api/stay/records":{"get":{"operationId":"StayController_findAllRecords","parameters":[],"responses":{"200":{"description":""}}}},"/api/stay/reserve/{id}":{"get":{"operationId":"StayController_reserve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/record":{"post":{"operationId":"RecordController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecordDto"}}}},"responses":{"201":{"description":""}}}},"/api/record/{type}":{"get":{"operationId":"RecordController_findAll","parameters":[{"name":"type","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/explore":{"post":{"operationId":"ExploreController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExploreDto"}}}},"responses":{"201":{"description":""}}}},"/api/explore/records":{"get":{"operationId":"ExploreController_findAllRecords","parameters":[],"responses":{"200":{"description":""}}}},"/api/explore/reserve/{id}":{"get":{"operationId":"ExploreController_reserve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/shop":{"post":{"operationId":"ShopController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateShopDto"}}}},"responses":{"201":{"description":""}}}},"/api/shop/records":{"get":{"operationId":"ShopController_findAllRecords","parameters":[],"responses":{"200":{"description":""}}}},"/api/shop/reserve/{id}":{"get":{"operationId":"ShopController_reserve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/eat":{"post":{"operationId":"EatController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEatDto"}}}},"responses":{"201":{"description":""}}}},"/api/eat/records":{"get":{"operationId":"EatController_findAllRecords","parameters":[],"responses":{"200":{"description":""}}}},"/api/eat/{meal}":{"get":{"operationId":"EatController_findAllMeal","parameters":[{"name":"meal","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/eat/reserve/{id}":{"get":{"operationId":"EatController_reserve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"cathygreen-api","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000","description":"Localhost"}],"components":{"schemas":{"VersionRes":{"properties":{"data":{"type":"object","properties":{"version":{"type":"string","readOnly":true,"example":"v1.0.0"}},"required":["version"]}},"required":["data"],"type":"object"},"CreateStayDto":{"type":"object","properties":{}},"CreateRecordDto":{"type":"object","properties":{}},"CreateExploreDto":{"type":"object","properties":{}},"CreateShopDto":{"type":"object","properties":{}},"CreateEatDto":{"type":"object","properties":{}}}}}